error handling and indentation fixes

+++ number-cruncher/src/Form.js
@@ -4,11 +4,11 @@ import React, { Component } from 'react';
 function printArray(a) {
     let s = "";
     for (let item of a) {
-      s += "["
-      for (let i = 0; i < item.length - 1; i++) {
-        s += item[i] + ", ";
-      }
-      s += item[item.length - 1] + "]"
+        s += "["
+        for (let i = 0; i < item.length - 1; i++) {
+            s += item[i] + ", ";
+        }
+        s += item[item.length - 1] + "]"
     }
     return s;
 };
@@ -20,10 +20,10 @@ function callFunction() {
     var arr = [];
     var arrStr = arrVal.split(",");
     for (var i = 0; i < arrStr.length; i++) {
-      arr.push(parseInt(arrStr[i]));
+        arr.push(parseFloat(arrStr[i]));
     }
   
-    var result = combinationSum(arr, parseInt(x));
+    var result = combinationSum(arr, parseFloat(x));
     var str = printArray(result);
   
     document.getElementById("result").innerHTML = str;
@@ -33,15 +33,15 @@ class Form extends Component {
     render() {
       return (
         <div className="App">
-          <h1>Number Cruncher</h1>
-          <form>
-            <label htmlFor="array">Number Array: </label>
-            <input type="text" id="array" name="array" /><br /><br />
-            <label htmlFor="target">Target: </label>
-            <input type="text" id="target" name="target" /><br /><br />
-            <input type="button" onClick={callFunction} value="Submit" />
-          </form>
-          <p id="result"></p>
+            <h1>Number Cruncher</h1>
+            <form>
+                <label htmlFor="array">Number Array: </label>
+                <input type="text" id="array" name="array" /><br /><br />
+                <label htmlFor="target">Target: </label>
+                <input type="text" id="target" name="target" /><br /><br />
+                <input type="button" onClick={callFunction} value="Submit" />
+            </form>
+            <p id="result"></p>
         </div>
       );
     }
